backend: create endpoint for downloading data created by cron
This solution raises a slight concern for the future - once we have a
lot of data collected it will be a while until the data downloads, thus
blocking the whole worker during this. IIRC we have 8 workers thus 8
downloads and API is unresponsive. This screams for async/await
solution - fastapi directly supports this (only adding async before endpoint,
the iterator is already written in this commit), but we will need to switch
to ASGI server - gunicorn to uvicorn.